Scenario of GHG (CO2eq) sequestration w/ seaweed
Taking the total GHG scenarios from Our Wolrd in Data, this file has the comparison of incrementally (by 0.15%) increasing sequestration via seaweed to essentially cover global GHG from agriculture (11-12%) by the end of the century. The csv provides the % increase of global GHG sequestered, how many tonnes that equates to, how much area (km2) of seaweed that would take, how much wet-weight production (tonnes), and how that compares to the BAU-current* modelled seaweed growth trend to understand how much of a gap needs to be filled to do this level of sequestration. There are values for the 'low' and 'high' emission scenarios, labeled as such.

*BAU-current exponential model based on 1990-2016 FAO global growth trends; R^2 = 98.5%).